job_search_automation/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3
lxml==4.9.3
webdriver-manager==4.0.1
urllib3==2.1.0
undetected-chromedriver==3.5.4
cloudscraper==1.2.71
fake-useragent==1.4.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
spacy==3.7.2
nltk==3.8.1
jsonlines==4.0.0
tabulate==0.9.0
aiohttp==3.9.1
asyncio-pool==0.6.0